feat: Make deny-blocked rules protocol and port specific
This change refactors the generation of  security rules to be specific
to the protocol and destination ports.

This prevents potential conflicts when multiple Services with different
port and protocol configurations share a single load balancer frontend.
By making the deny rules more granular, we ensure that traffic is only
blocked for the intended service, protocol, and port combination.