feat: Ensure correct NSG rule ordering
Ensure correct NSG rule ordering between explicit allow and explicit
deny rules.

In order to maintain consistent semantics, the explicit prefix deny
and explicit allow rules cannot randomly interleave. The deny rules
must also take precedence over the allow NSG rules.

This changes the rule priority generation logic to allocate the deny
rules from the start of the reserved NSG ID range and allow rules from
the end of the range.

A small ID space is reserved at the end of the range for the "deny all"
rules.

The code assumes the allow and deny rules never interleave due
to the quota limitation on the total number of rules in an NSG (1000).